Function to method, update doc, warn about caching
